+++ custom_components/kodi_media_sensors/entity_kodi_media_sensor.py
@@ -1,10 +1,13 @@
+from asyncio import FastChildWatcher
 import logging
 import json
 from datetime import datetime
 from typing import Any, Dict, List, Optional
 from pykodi import Kodi
 from urllib import parse
+from config.custom_components.media_sensor_event_manager import MediaSensorEventManager
 from homeassistant.helpers.entity import Entity
+from homeassistant import core
 from homeassistant.const import (
     STATE_OFF,
     STATE_ON,
@@ -34,14 +37,20 @@ class KodiMediaSensorEntity(Entity, ABC):
 
     def __init__(
         self,
+        hass: core.HomeAssistant,
         kodi: Kodi,
         config: KodiConfig,
-        hide_watched: bool = False,
+        event_manager: MediaSensorEventManager,
+        # hide_watched: bool = False,
     ) -> None:
         super().__init__()
+        self._hass = hass
         self._kodi = kodi
         self._define_base_url(config)
+        self._event_manager = event_manager
         self._state = STATE_OFF
+
+        self._event_manager.register_sensor(self)
 
     def _define_base_url(self, config):
         protocol = "https" if config["ssl"] else "http"
@@ -51,6 +60,10 @@ class KodiMediaSensorEntity(Entity, ABC):
         self._base_web_url = (
             f"{protocol}://{auth}{config['host']}:{config['port']}/image/image%3A%2F%2F"
         )
+
+    @abstractmethod
+    async def handle_media_sensor_event(self, event):
+        _LOGGER.warning("This method is not implemented for the entity")
 
     @abstractmethod
     async def async_call_method(self, method, **kwargs):
@@ -99,7 +112,7 @@ class KodiMediaSensorEntity(Entity, ABC):
                 new_data: List[Dict[str, Any]] = result.get(entry, [])
                 default_type = MAP_KEY_MEDIA_TYPE.get(entry)
 
-                if self._hasLeaf(default_type):
+                if self._has_leaf(default_type):
                     for item in new_data:
                         self._format_item(item, default_type)
                 else:
@@ -107,7 +120,7 @@ class KodiMediaSensorEntity(Entity, ABC):
 
         return new_data
 
-    def _hasLeaf(self, default_type):
+    def _has_leaf(self, default_type):
         if (
             default_type == MEDIA_TYPE_ALBUM_DETAIL
             or default_type == MEDIA_TYPE_TVSHOW_DETAIL
@@ -123,8 +136,8 @@ class KodiMediaSensorEntity(Entity, ABC):
             item["genre"] = ", ".join(item["genre"])
 
         if "thumbnail" in item:
-            th = item["thumbnail"]
-            if th is None or th == "":
+            thumbnail = item["thumbnail"]
+            if thumbnail is None or thumbnail == "":
                 del item["thumbnail"]
             else:
                 thumbnail = self._kodi.thumbnail_url(item["thumbnail"])
@@ -201,9 +214,9 @@ class KodiMediaSensorEntity(Entity, ABC):
         self._attrs["data"] = json.dumps(self._data)
 
     def init_meta(self, event_id):
-        ds = datetime.now().strftime(UPDATE_FORMAT)
+        formatted_date = datetime.now().strftime(UPDATE_FORMAT)
         self.purge_meta(event_id)
-        self._meta[0]["update_time"] = ds
+        self._meta[0]["update_time"] = formatted_date
         self._meta[0]["sensor_entity_id"] = self.domain_unique_id
         self._meta[0]["service_domain"] = DOMAIN
         self.build_attrs()